\name{LINselect-package}
\alias{LINselect-package}
\alias{LINselect}
\docType{package}
\title{Selection of linear estimators}
\description{LINselect allows to estimate the mean of a Gaussian vector, by choosing among a large collection of estimators. In particular it solves the problem of variable selection by choosing the best predictor among  predictors emanating from different methods as lasso, elastic-net, adaptive lasso, pls, randomForest. Moreover, it can be applied for choosing the tuning parameter in a Gauss-lasso procedure.}
\details{
\tabular{ll}{Package: \tab LINselect\cr
Title: \tab Selection of linear estimators\cr
Version: \tab 0.0-0\cr
Date: \tab 2013-03-01\cr
Author: \tab Yannick Baraud, Christophe Giraud, Sylvie Huet\cr
Maintainer: \tab Annie Bouvier <Annie.Bouvier@jouy.inra.fr>\cr
Suggests: \tab mvtnorm, elasticnet, MASS, randomForest, pls, gtools\cr
License: \tab GPL (>= 3)\cr
URL: \tab \cr
Encoding: \tab latin1\cr}
}
\author{Yannick Baraud, Christophe Giraud, Sylvie Huet}

\keyword{ package }


